+++ src/Cli/SubCommands/HelpCommand.ts
@@ -1,13 +1,6 @@
 import Command from "./Command"
 import chalk from "chalk"
-import stripAnsi from 'strip-ansi'
-
-const WIDTH = 70
-
-const logOnBlack = (message: string = '') => {
-    const offset = message.length - stripAnsi(message).length
-    console.log(chalk.bgHex('#000000')(message.padEnd(WIDTH + offset)))
-}
+import logOnBlack from "../Output/logOnBlack"
 
 const HelpCommand: Command = {
     name: 'help',
@@ -15,13 +8,17 @@ const HelpCommand: Command = {
     run: async () => {
         console.log()
         logOnBlack()
-        logOnBlack(chalk.bgHex('#000000')(`    ${chalk.hex('#FEC908')('J E ')}${chalk.hex('#F9A619')('T P ')}${chalk.hex('#F25822')('A X ')}`))
+        logOnBlack(`    ${chalk.hex('#FEC908')('J E ')}${chalk.hex('#F9A619')('T P ')}${chalk.hex('#F25822')('A X ')}`)
         logOnBlack(chalk.hex('#FFFFFF')(`    C O M M A N D S`))
         logOnBlack()
         logOnBlack(`    ${chalk.hex('#FEC908')('up')}           ${chalk.hex('#F9A619')(`Starts the Jetpax server`)}`)
         logOnBlack(`    ${chalk.hex('#FEC908')('down')}         ${chalk.hex('#F9A619')(`Stops the Jetpax server`)}`)
         logOnBlack(`    ${chalk.hex('#FEC908')('status')}       ${chalk.hex('#F9A619')(`Displays the status of the Jetpax server`)}`)
         logOnBlack(`    ${chalk.hex('#FEC908')('logs')}         ${chalk.hex('#F9A619')(`Tails the Jetpax server logs`)}`)
+        logOnBlack()
+        logOnBlack(`    ${chalk.hex('#FEC908')('init')}         ${chalk.hex('#F9A619')(`Initializes a workspace (in the current directory)`)}`)
+        logOnBlack(`    ${chalk.hex('#FEC908')('register')}     ${chalk.hex('#F9A619')(`Registers a workspace `)}`)
+        logOnBlack(`    ${chalk.hex('#FEC908')('deregister')}   ${chalk.hex('#F9A619')(`Deregisters a workspace`)}`)
         logOnBlack()
         logOnBlack(`    ${chalk.hex('#FEC908')('ui')}           ${chalk.hex('#F9A619')(`Opens the Jetpax dashboard`)}`)
         logOnBlack(`    ${chalk.hex('#FEC908')('web')}          ${chalk.hex('#F9A619')(`Opens the Jetpax dashboard in a browser`)}`)